Fix installation under Ubuntu Python
On Ubuntu, distutils has been patched[1] to, by default, use a different installation layout than is standard on Unix
systems (to more closely follow FHS). As such, using sys.prefix to retrieve the location where data files are placed is
not correct. This leads to installation failing. To get around this, use the resource_filename[2] method provided by
setuptools, which returns the correct location regardless of the platform.

[1] http://bazaar.launchpad.net/~ubuntu-branches/ubuntu/precise/python2.7/precise/view/head:/debian/patches/distutils-install-layout.diff
[2] http://peak.telecommunity.com/DevCenter/PythonEggs#accessing-package-resources