Make python tests pass.
Have python client support v2_replacement and os_v1 and os_v2 replacement.

Python client uses the None value instead of 'Other' when the device isn't found. Make the test reflect that.

The probable correct fix is to make the parse return 'Other' in that case, but this would be a change to the behavior that others are probably depending on.