GoogleTV "old-style" version may contain non digit characters
GoogleTV user-agents may end up with tokens such as "GoogleTV/b39389"
(the character of interest is the "b" in "b39389"). The parser that
detects the GoogleTV OS rather expects the version to be composed of
digits.

I replaced "\d+" with "[\da-z]+". This could perhaps be further
restricted as I have never seen examples of model versions that
contain but do not start with a non digit character, but I doubt
being a tad more flexible creates any false positive.

Test added along other GoogleTV tests in additional_os_tests.yaml