Clarify copyright status of the project(s)
The readme notes a lot of different copyrights and licensing agreements. While the licensing is great, the components are no longer copyright [originator], they're a derivative work consisting of copyrights held by all the contributors _to_ the YAML file, or the port, or etc, etc, etc. 

As a result, the readme is inaccurately representing the licensing around the project. This change should fix that, and do so in a way that doesn't require updating - although it will necessitate some changes if/when we switch to using distinct repositories under a single org banner.